apps/api-server/pyproject.toml
[project]
name = "mirothinker-api-server"
version = "0.1.0"
description = "MiroThinker API Server - OpenAI Compatible Research Agent API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "miroflow-agent",
    "sse-starlette>=2.2.1",
    "openai>=1.78.1",
    "tiktoken>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["routers", "services"]
include = ["main.py"]

[tool.uv.sources]
miroflow-agent = { path = "../miroflow-agent", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
]

